+++ packages/api/src/deleteCertification.ts
@@ -1,4 +1,4 @@
-import instance from "../axios";
+import instance from "./axios";
 
 export const removeCertification = async (id: string | number) => {
   return instance.delete(`/certificates/current/${id}`);
